edit mbti_pthread - append comment & bug fix
