+++ mbti/mbti_pthread/mbti_pthread_global_attr.c
@@ -1,0 +1,2 @@
+#include "mbti_pthread_global.h"
+    unsigned int pthread_g_counter;
